Разбил Js на модули и добавил пустые файлы, под будущие модули
