+++ js/main.js
@@ -1,2 +1,5 @@
-import './data.js';
+import {getPhotosArray} from './data.js';
 import './util.js';
+import './comments.js';
+
+getPhotosArray(25);
